For fish
ADD THE FOLLOWING INTO YOUR ANDROID MANIFEST
<uses-permission android:name="android.permission.CAMERA"/>
this one allows you to send a thumbnail over. if you want to combine with the sending bit later, you need to look at MainActivity.java (mine) to get the method.
Another thing: Camera is kinda buggy in emulator. But this should work (it worked last time for me)